# backend/internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  pointmarket_backend_internal_dtos.APIError:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  pointmarket_backend_internal_dtos.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  pointmarket_backend_internal_dtos.CreateMissionRequestDTO:
    properties:
      description:
        type: string
      metadata:
        items:
          type: integer
        type: array
      reward_points:
        type: integer
      title:
        type: string
    required:
    - metadata
    - title
    type: object
...
  pointmarket_backend_internal_dtos.CreateUserMissionRequestDTO:
    properties:
      mission_id:
        type: integer
      progress:
        items:
          type: integer
        type: array
      status:
        description: e.g., "not_started", "in_progress"
        type: string
      user_id:
        type: integer
    required:
    - mission_id
    - user_id
    type: object
...
  pointmarket_backend_internal_dtos.ListUserMissionsResponseDTO:
    properties:
      total:
        type: integer
      user_missions:
        items:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.UserMissionDTO'
        type: array
    type: object
...
  pointmarket_backend_internal_dtos.MissionDTO:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      metadata:
        items:
          type: integer
        type: array
      reward_points:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  pointmarket_backend_internal_dtos.NullData:
    type: object
...
  pointmarket_backend_internal_dtos.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/pointmarket_backend_internal_dtos.PaginationMeta'
      success:
        type: boolean
    type: object
  pointmarket_backend_internal_dtos.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  pointmarket_backend_internal_dtos.UpdateMissionRequestDTO:
    properties:
      description:
        type: string
      metadata:
        items:
          type: integer
        type: array
      reward_points:
        type: integer
      title:
        type: string
    type: object
...
  pointmarket_backend_internal_dtos.UpdateUserMissionStatusRequestDTO:
    properties:
      completed_at:
        description: Optional, set if status is completed
        type: string
      status:
        description: e.g., "completed", "in_progress"
        type: string
    required:
    - status
    type: object
...
  pointmarket_backend_internal_dtos.UserMissionDTO:
    properties:
      completed_at:
        type: string
      id:
        type: integer
      mission_description:
        type: string
      mission_id:
        type: integer
      mission_metadata:
        items:
          type: integer
        type: array
      mission_reward_points:
        type: integer
      mission_title:
        type: string
      progress:
        items:
          type: integer
        type: array
      started_at:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
  pointmarket_backend_internal_dtos.UserMissionProgressDTO:
    properties:
      completed_at:
        type: string
      mission_title:
        type: string
      started_at:
        type: string
      status:
        type: string
      user_email:
        type: string
      user_name:
        type: string
    type: object
...
info:
  contact: {}
  description: PointMarket backend API documentation.
  title: PointMarket API
  version: "1.0"
paths:
...
  /missions:
    get:
      description: Default response returns all missions. If `user_id` is provided,
        returns all missions for that user (admin or owner).
      parameters:
      - description: User ID (admin or owner)
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.ListUserMissionsResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: List missions
      tags:
      - missions
    post:
      consumes:
      - application/json
      parameters:
      - description: Mission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.CreateMissionRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.MissionDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Create mission
      tags:
      - missions
  /missions/{id}:
    delete:
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.NullData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Delete mission
      tags:
      - missions
    get:
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.MissionDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get mission by ID
      tags:
      - missions
    put:
      consumes:
      - application/json
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Mission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.UpdateMissionRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.MissionDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Update mission
      tags:
      - missions
  /missions/{id}/end:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      - description: User mission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.CreateUserMissionRequestDTO'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: End mission (not implemented)
      tags:
      - missions
  /missions/{id}/start:
    post:
      consumes:
      - application/json
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      - description: User mission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.CreateUserMissionRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.UserMissionDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Start mission
      tags:
      - missions
  /missions/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: User mission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.UpdateUserMissionStatusRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.UserMissionDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Update user mission status
      tags:
      - missions
  /missions/progress:
    get:
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pointmarket_backend_internal_dtos.UserMissionProgressDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: List missions progress
      tags:
      - missions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"